refactor: extract Declined handler from handleStateTransitions
Move the Declined branch into effects/declinedEffects.ts (~170 lines
— the largest single-state handler). Handles:
- Stamping declinedAt / declinedBy onto firestoreUpdates
- Composing and sending the decline email with tenant grace period
  and per-service decline detail (for parallel-state declines)
- PUTting /api/calendarEvents to prefix the Google Calendar title
  with [DECLINED]

HandlerContext gains currentSnapshot and newSnapshot fields since
the decline email composition reads newSnapshot.context for
servicesApproved / servicesRequested / declineReason. Other extracted
handlers are unaffected — the new fields are optional reads.

No behavior change. tsc clean, 44 tests pass.
